feat: export AssertionError and ValidationError
Add ValidationError to help distinguish between unrecoverable errors
while decoding an `.sb` file and encoding to Scratch 2 data.

Export AssertionError and ValidationError so users of the converter can
know when an error is from validating the Scratch 1 file (as in it
likely is not an Scratch 1 file) or from making sure the Scratch 1 file
data matches expected ranges (like not making memory buffers that are
too big).